[github-348] tidy up number casting
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1902803 13f79535-47bb-0310-9956-ffa450edef68

+++ poi-ooxml/src/main/java/org/apache/poi/xssf/binary/XSSFBParser.java
@@ -92,7 +92,7 @@ public abstract class XSSFBParser {
             b2 &= ~(1<<7); //unset highest bit (if it exists?)
             recordId = ((int)b2 << 7)+(int)b1;
         } else {
-            recordId = (int)b1;
+            recordId = b1;
         }
 
         long recordLength = 0;
@@ -102,7 +102,7 @@ public abstract class XSSFBParser {
             byte b = is.readByte();
             halt = (b >> 7 & 1) == 0; //if highest bit !=1 then continue
             b &= ~(1<<7);
-            recordLength += (int)b << (i*7); //multiply by 128^i
+            recordLength += (long) (int)b << (i*7); //multiply by 128^i
             i++;
 
         }
